// src/api/analytics/model/modelDriverAverageTimeResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * TL Analytics API
 * API для аналитики заказов
 * OpenAPI spec version: 1.0
 */

export interface ModelDriverAverageTimeResponse {
  /** среднее время в минутах */
  averageTimeMinutes?: number;
  driverName?: string;
}