// src/api/analytics/model/modelDriverOrdersResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * TL Analytics API
 * API для аналитики заказов
 * OpenAPI spec version: 1.0
 */
import type { ModelDriverOrdersResponseCargoTypes } from './modelDriverOrdersResponseCargoTypes';

export interface ModelDriverOrdersResponse {
  /** float для веса, int для количества */
  cargoTypes?: ModelDriverOrdersResponseCargoTypes;
  driverName?: string;
}