// src/api/analytics/model/modelDepartIncomingResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * TL Analytics API
 * API для аналитики заказов
 * OpenAPI spec version: 1.0
 */
import type { ModelDepartIncomingResponseCargoTypes } from './modelDepartIncomingResponseCargoTypes';

export interface ModelDepartIncomingResponse {
  cargoTypes?: ModelDepartIncomingResponseCargoTypes;
  departName?: string;
}